+++ tests/testthat/test-all.R
@@ -1,4 +1,22 @@
 test_that(
+  "Column labels are shown as names in data listings if present" |>
+    vdoc[["add_spec"]](c(specs$listings$column_labels)),
+  {
+    app <- shinytest2::AppDriver$new(root_app_url)
+
+    app$wait_for_js("document.querySelector('#papo-listings-column_selector_adae') !== null")
+    app$set_inputs("papo-listings-column_selector_adae" = c("USUBJID"))
+
+    app$wait_for_value(export = "papo-listings-test_data")
+    selected_data <- app$get_value(export = "papo-listings-test_data")[["filtered_data"]]
+
+    expect_equal(attr(selected_data[["USUBJID"]], "label"), "Unique Subject Identifier")
+
+    app$stop()
+  }
+)
+
+test_that(
   "functional subject_selector" |>
     vdoc[["add_spec"]](c(specs$common$subject_selector)),
   {
@@ -38,25 +56,28 @@ test_that(
     vdoc[["add_spec"]](c(specs$common$bookmarking)),
   {
     app <- shinytest2::AppDriver$new(root_app_url)
-    app$wait_for_idle(wait_for_idle_ms)
 
     sel_id <- "papo-patient_selector"
+    test_pid <- "01-701-1028"
+    default_pid <- "01-701-1015"
 
-    inputs <- list()
-    inputs[[sel_id]] <- "01-701-1028"
+    app$wait_for_value(input = sel_id)
 
-    do.call(app$set_inputs, inputs)
-    app$wait_for_idle(wait_for_idle_ms)
+    app$set_inputs(!!sel_id := test_pid)
+    app$wait_for_value(input = sel_id, ignore = list(NULL, "", "01-701-1015"))
 
     bmk_url <- app$get_js("window.location.href")
-
     bookmark_app <- shinytest2::AppDriver$new(bmk_url)
-    bookmark_app$wait_for_idle(wait_for_idle_ms)
-    app_input_values <- app$get_values()[["input"]]
-    bmk_input_values <- bookmark_app$get_values()[["input"]]
+    bookmark_app$wait_for_value(input = sel_id, ignore = list(NULL, "", "01-701-1015"))
 
-    expect_equal(app_input_values[[sel_id]], bmk_input_values[[sel_id]])
+    app_input_value <- app$get_value(input = sel_id)
+    bmk_input_value <- bookmark_app$get_value(input = sel_id)
+
+    expect_equal(app_input_value, test_pid)
+    expect_equal(bmk_input_value, test_pid)
+
     app$stop()
+    bookmark_app$stop()
   }
 )
 
@@ -65,12 +86,18 @@ test_that(
     vdoc[["add_spec"]](c(specs$common$jump_to_subject)),
   {
     app <- shinytest2::AppDriver$new(root_app_url)
-    app$wait_for_idle(wait_for_idle_ms)
+
+    sel_id <- "papo-patient_selector"
+
+    app$wait_for_value(input = sel_id)
+
     app$click(input = "jump")
-    app$wait_for_idle(wait_for_idle_ms)
+    app$wait_for_value(input = sel_id, ignore = list(NULL, "", "01-701-1015"))
+
     expected <- "01-701-1033"
-    actual <- app$get_value(input = "papo-patient_selector")
+    actual <- app$get_value(input = sel_id)
     expect_equal(actual, expected)
+
     app$stop()
   }
 )
@@ -84,13 +111,16 @@ test_that(
       name = "misconfigured_app"
     )
 
-    app$wait_for_idle(wait_for_idle_ms)
+    app$wait_for_js(
+      "document.querySelector('#papo-validator-ui').innerText.includes('missing')",
+      timeout = 30000
+    )
 
     validation_errors <- app$get_html(selector = "#papo-validator-ui")
-    expect_true(grepl("`subject_level_dataset_name` missing", validation_errors, fixed = TRUE))
-    expect_true(grepl("`subjid_var` missing", validation_errors, fixed = TRUE))
-    expect_true(grepl("The `sender_ids` - 'random1' - are not available.",
-                      validation_errors, fixed = TRUE))
+
+    expect_match(validation_errors, "`subject_level_dataset_name` missing", fixed = TRUE)
+    expect_match(validation_errors, "`subjid_var` missing", fixed = TRUE)
+    expect_match(validation_errors, "The `sender_ids` - 'random1' - are not available.", fixed = TRUE)
 
     app$stop()
   }
@@ -105,22 +135,31 @@ test_that(
       name = "patient_selector_switching"
     )
 
-    app$set_inputs(`global_filter-vars` = "SEX")
-    app$wait_for_idle()
-
-    # Check if the first patient was selected (initial state)
-    testthat::expect_equal(app$get_value(input = "mock_app-patient_selector"), "01-701-1015")
+    sel_id <- "mock_app-patient_selector"
 
     # Check if the first male patient was selected when filtered accordingly
-    app$set_inputs(`global_filter-SEX` = "M")
-    # assign pat_id only to suppress print
-    pat_id <- app$wait_for_value(input = "mock_app-patient_selector", ignore = list("01-701-1015"))
-    testthat::expect_equal(app$get_value(input = "mock_app-patient_selector"), "01-701-1023")
+    app$set_inputs(`filter-filter_state_json_input` = paste0(
+      '{"filters":{"datasets_filter":{"children":[]},',
+      '"subject_filter":{"children":[{"kind":"row_operation",',
+      '"operation":"and","children":[{"kind":"filter",',
+      '"dataset":"adsl","operation":"select_subset",',
+      '"variable":"SEX","values":["M"],"include_NA":true}]}]}},',
+      '"dataset_list_name":"demo"}'
+    ), allow_no_input_binding_ = TRUE, priority_ = "event")
+    pat_id <- app$wait_for_value(input = sel_id)
+    testthat::expect_equal(app$get_value(input = sel_id), "01-701-1023")
 
     # Check if no patient is selected when filtered accordingly
-    app$set_inputs(`global_filter-SEX` = character(0))
-    app$wait_for_idle(2000)
-    testthat::expect_equal(app$get_value(input = "mock_app-patient_selector"), "")
+    app$set_inputs(`filter-filter_state_json_input` = paste0(
+      '{"filters":{"datasets_filter":{"children":[]},',
+      '"subject_filter":{"children":[{"kind":"row_operation",',
+      '"operation":"and","children":[{"kind":"filter",',
+      '"dataset":"adsl","operation":"select_subset",',
+      '"variable":"SEX","values":[],"include_NA":true}]}]}},',
+      '"dataset_list_name":"demo"}'
+    ), allow_no_input_binding_ = TRUE, priority_ = "event")
+    pat_id <- app$wait_for_value(input = sel_id, ignore = list(NULL, "01-701-1023"))
+    testthat::expect_equal(app$get_value(input = sel_id), "")
 
     app$stop()
   }
@@ -160,27 +199,12 @@ test_that(
 )
 
 test_that(
-  "Column labels are shown as names in data listings if present" |>
-    vdoc[["add_spec"]](c(specs$listings$column_labels)),
-  {
-    app <- shinytest2::AppDriver$new(root_app_url)
-    app$wait_for_idle(wait_for_idle_ms)
-    app$set_inputs("papo-listings-column_selector_adae" = c("USUBJID"))
-    app$wait_for_idle(wait_for_idle_ms)
-
-    selected_data <- app$get_values()[["export"]][["papo-listings-test_data"]][["filtered_data"]]
-    expect_equal(attr(selected_data[["USUBJID"]], "label"), "Unique Subject Identifier")
-
-    app$stop()
-  }
-)
-
-test_that(
   "Tab controls allow listing switching" |>
     vdoc[["add_spec"]](c(specs$listings$switching, specs$listings$extra_column_selection)),
   {
     app <- shinytest2::AppDriver$new(root_app_url)
     app$wait_for_idle(duration = wait_for_idle_ms)
+
     app$set_inputs("papo-listings-data_selector" = "cm")
     app$wait_for_idle()
 
@@ -220,7 +244,7 @@ test_that(
     testthat::expect_equal(listing_properties[["filter"]], "top")
     testthat::expect_equal(listing_properties[["options"]][["searching"]], TRUE)
     testthat::expect_equal(listing_properties[["options"]][["ordering"]], TRUE)
-    testthat::expect_equal(listing_properties[["options"]][["buttons"]][["text"]], "Reset Columns Order")
+    testthat::expect_equal(listing_properties[["options"]][["buttons"]][["text"]], "Reset Rows Order")
 
     filter_html <- listing_properties[["filterHTML"]]
     testthat::expect_true(grepl("select multiple", filter_html) && grepl('data-type="factor"', filter_html))
@@ -279,12 +303,11 @@ test_that(
     app <- shinytest2::AppDriver$new(root_app_url)
     app$wait_for_idle(duration = wait_for_idle_ms)
 
-    target_color <- grDevices::col2rgb(CONST$default_palette[["MILD"]])
-    target_color_as_string <- paste0("rgba(", paste(target_color, collapse = ","), ",1)")
+    target_color <- CONST$default_palette[["MILD"]]
 
     ae_plot_first_color <-
       app$get_values()[["export"]][["papo-plot_contents-test_plot_data"]][["plot_first_line_color/Adverse Events Plot"]]
-    expect_equal(target_color_as_string, ae_plot_first_color)
+    expect_equal(target_color, ae_plot_first_color)
 
     app$stop()
   }
@@ -296,11 +319,25 @@ test_that(
   {
     app <- shinytest2::AppDriver$new(root_app_url)
     app$wait_for_idle(wait_for_idle_ms)
-    app$click(input = "jump")
+
+    app$set_inputs(`papo-patient_selector` = "01-701-1429")
     app$wait_for_idle(duration = wait_for_idle_ms)
-    expected <- "01-701-1033"
+
     plot_messages <- app$get_values()[["export"]][["papo-plot_contents-test_plot_data"]][["plot_messages"]]
     expect_contains(plot_messages, "* No Data for Adverse Events Plot.")
+    expect_contains(plot_messages, "* No Data for Concomitant Medication Plot.")
+    expect_no_match(plot_messages, "\\* No Parameter for Lab plot selected\\.")
+    expect_no_match(plot_messages, "\\* No Parameter for Vital Sign Plot selected\\.")
+
+    # Deselect all lab and vitals parameters
+    app$set_inputs(`papo-plot_contents-Labplot` = character(0),
+                   `papo-plot_contents-VitalSignPlot` = character(0))
+    app$wait_for_idle(duration = wait_for_idle_ms)
+
+    plot_messages <- app$get_values()[["export"]][["papo-plot_contents-test_plot_data"]][["plot_messages"]]
+    expect_contains(plot_messages, "* No Parameter for Lab plot selected.")
+    expect_contains(plot_messages, "* No Parameter for Vital Sign Plot selected.")
+
     app$stop()
   }
 )
@@ -405,6 +442,7 @@ test_that(
       app_dir = "apps/grading_palette_colors/",
       name = "grading_colors_app"
     )
+    app$wait_for_idle(wait_for_idle_ms)
 
     app_grading_vals <- setdiff(app$get_value(export = "gradings"), NA)
     app_filled_palette <- app$get_value(export = "filled_palette")
@@ -421,5 +459,6 @@ test_that(
     # ii. check a color was then assigned.
     expect_length(grading_palette[unmapped_grading_vals], length(unmapped_grading_vals))
 
+    app$stop()
   }
 )